refactor: extract kodi checks in own module
Package together what changes together.
